- HomeLink storage is now Seed based.
To restore the previously saved homes:
- Load this version into your Server and start the server. This version will automatically generate a folder that has the seed in the name.
- Stop the server and copy the .json files containing players home locations from HomeLink folder into the newly created HomeLink folder that has the server Seed in the name.
- You can safely delete the empty folder and then start the server.

+++ src/main/java/com/sumutiu/homelink/commands/DelHomeCommand.java
@@ -3,10 +3,12 @@ package com.sumutiu.homelink.commands;
 import com.mojang.brigadier.CommandDispatcher;
 import com.mojang.brigadier.arguments.StringArgumentType;
 import com.sumutiu.homelink.storage.HomeStorage;
-import com.sumutiu.homelink.util.HomeLinkMessages;
 import net.minecraft.commands.CommandSourceStack;
 import net.minecraft.commands.Commands;
 import net.minecraft.server.level.ServerPlayer;
+
+import static com.sumutiu.homelink.HomeLink.HomeLinkInitialized;
+import static com.sumutiu.homelink.util.HomeLinkMessages.*;
 
 public class DelHomeCommand {
 
@@ -18,34 +20,29 @@ public class DelHomeCommand {
                                 .executes(ctx -> {
 
                                     CommandSourceStack source = ctx.getSource();
-
                                     if (!(source.getEntity() instanceof ServerPlayer player)) {
-                                        HomeLinkMessages.Logger(1, HomeLinkMessages.PLAYER_ONLY_COMMAND);
+                                        Logger(1, PLAYER_ONLY_COMMAND);
                                         return 0;
                                     }
 
-                                    String name = StringArgumentType.getString(ctx, "name");
+                                    if (HomeLinkInitialized) {
+                                        String name = StringArgumentType.getString(ctx, "name");
 
-                                    if (HomeStorage.getHome(player, name) != null) {
+                                        if (HomeStorage.getHome(player, name) != null) {
+                                            HomeStorage.deleteHome(player, name);
+                                            PrivateMessage(player, String.format(HOME_DELETED, name));
 
-                                        HomeStorage.deleteHome(player, name);
+                                            return 1;
+                                        } else {
+                                            PrivateMessage(player, String.format(HOME_NOT_FOUND, name));
 
-                                        HomeLinkMessages.PrivateMessage(
-                                                player,
-                                                String.format(HomeLinkMessages.HOME_DELETED, name)
-                                        );
-
-                                        return 1;
-
+                                            return 0;
+                                        }
                                     } else {
-
-                                        HomeLinkMessages.PrivateMessage(
-                                                player,
-                                                String.format(HomeLinkMessages.HOME_NOT_FOUND, name)
-                                        );
-
+                                        PrivateMessage(player, MOD_INIT_NOT_READY);
                                         return 0;
                                     }
+
                                 })
                         )
         );
